requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
Flask==2.1.1
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
lxml==4.8.0
MarkupSafe==2.1.1
Pillow==9.1.0
pip-autoremove==0.10.0
psycopg2-binary==2.9.3
python-docx==0.8.11
python-dotenv==0.20.0
requests==2.27.1
urllib3==1.26.9
Werkzeug==2.1.1
